Crate system_tracing

System Tracing Module

System Tracing crate helps you dynamically trace instructions and functions in both user mode and kernel mode of the operating system via addresses, and allows for custom-designed tracing code.

Enums

ProbePlace
Store the location of the functions that need to be traced, either in user mode or kernel mode, which includes ['ProbeType'].

ProbeType
Store the type of tracing required, including instruction tracing, synchronous function tracing, and asynchronous function tracing (not implemented).

Functions

kprobe_register
Kprobes Register

kprobe_unregister

kprobes_trap_handler
Kprobes Trap Handler

uprobe_register
Uprobes Register

uprobes_init
Uprobes Init

uprobes_kernel_function_initialization
Uprobes Kernel Function Initialization

uprobes_trap_handler
Uprobes Trap Handler