Function system_tracing::kprobe_register

pub fn kprobe_register(
    addr: usize, 
    handler: Arc<Mutex<dyn FnMut(&mut TrapFrame) + Send>>, 
    post_handler: Option<Arc<Mutex<dyn FnMut(&mut TrapFrame) + Send>>>, 
    probe_type: ProbeType
) -> isize

Kprobes Register

You can use this function to register trace points in kernel mode. You need to provide the address to be traced, a handler function to be executed before the instruction/function runs, an optional handler function to be executed after the instruction/function has completed, and ['ProbeType'].

Example

system_tracing::kprobe_register(
    self.addr,
    alloc::sync::Arc::new(Mutex::new(move |cx: &mut UserContext| {
        interpret(&prog, &HELPERS, cx as *const UserContext as usize as u64);
    })),
    None))),
    ProbeType::SyncFunc
)