Function system_tracing::kprobes_trap_handler

pub fn kprobes_trap_handler(cx: &mut TrapFrame)

Kprobes Trap Handler

If you want to use Kprobes you need to place it in the operating system's handler function for the ebreak interrupt occurring in kernel mode.

The type of the parameter is &mut trapframe::TrapFrame

[dependencies]
trapframe = { git = "https://github.com/rcore-os/trapframe-rs", rev = "bdfe5aa" }

Example

pub fn trap_handler_no_frame(tf: &mut TrapFrame) {
    match scause::read().cause() {
        Trap::Exception(E::Breakpoint) => { kprobes_trap_handler(tf) }
    }
}