Function system_tracing::uprobe_register

pub fn uprobe_register(
    path: String, 
    addr: usize, 
    handler: Arc<Mutex<dyn FnMut(&mut UserContext) + Send>>, 
    post_handler: Option<Arc<Mutex<dyn FnMut(&mut UserContext) + Send>>>, 
    probe_type: ProbeType
) -> isize

Uprobes Register

You can use this function to register trace points in user mode. You need to provide the path of the user-mode program to be traced, the address within the user-mode program that needs to be traced, a handler function to be executed before the instruction/function runs, an optional handler function to be executed after the instruction/function has completed, and ['ProbeType'].

Example

system_tracing::uprobe_register(
    "rust/test_uprobes",
    self.addr,
    alloc::sync::Arc::new(Mutex::new(move |cx: &mut UserContext| {
        interpret(&prog, &HELPERS, cx as *const UserContext as usize as u64);
    })),
    Some(alloc::sync::Arc::new(Mutex::new(move |cx: &mut UserContext| {
        test_post_handler(cx);
    }))),
    ProbeType::Insn
)