Function system_tracing::uprobes_kernel_function_initialization

pub fn uprobes_kernel_function_initialization(
    get_exec_path: Arc<Mutex<dyn FnMut() -> String + Send>>, 
    get_new_page: Arc<Mutex<dyn FnMut(usize, usize) -> usize + Send>>
)

Uprobes Kernel Function Initialization

If you want to use Uprobes, you need to first initialize a add_new_page function and a 'get_exec_path' function through the init function, before user mode starts.

Function feature requirements：

add new page

Given an address and a required length, the function creates a blank page closest to the address and returns the address of the blank page.

get execution path

Get the path of the currently running user-mode program.

Example

fn get_exec_path() -> String{
    current_thread().unwrap().proc.try_lock().expect("locked!").exec_path.clone()
}

fn get_new_page(addr: usize, len: usize) -> usize{
    let thread = current_thread().unwrap();
    let mut vm = thread.vm.lock();
    let addr = vm.find_free_area(addr, len);
    vm.push(
        addr,
        addr + len,
        MemoryAttr::default().user().execute().writable(),
        ByFrame::new(GlobalFrameAlloc),
        "point",
    );
    unsafe {asm!("fence.i");}
    addr
}

pub fn init(){
    system_tracing::uprobes_kernel_function_initialization(
        Arc::new(Mutex::new(|| get_exec_path())),
        Arc::new(Mutex::new(move |addr: usize, len: usize| get_new_page(addr, len)))
    )
}