Function system_tracing::uprobes_trap_handler

pub fn uprobes_trap_handler(cx: &mut UserContext)

Uprobes Trap Handler

If you want to use Uprobes you need to place it in the operating system's handler function for the ebreak interrupt occurring in user mode.

The type of the parameter is &mut trapframe::UserContext

[dependencies]
trapframe = { git = "https://github.com/rcore-os/trapframe-rs", rev = "bdfe5aa" }

Example

match trap_num {
	_ if is_ebreak(trap_num) => {
        let cx: &mut UserContext = &mut thread_context.user;
        systemp_tracing::uprobes_trap_handler(cx);
    }
}